Add proposal loader dapp
