+++ sample/dapp/src/main/kotlin/com/walletconnect/sample/dapp/ui/routes/composable_routes/chain_selection/ChainSelectionRoute.kt
@@ -7,6 +7,9 @@ import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.BoxScope
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
@@ -14,25 +17,31 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.CircularProgressIndicator
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.vectorResource
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.tooling.preview.PreviewParameterProvider
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
 import androidx.core.net.toUri
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavController
@@ -44,26 +53,26 @@ import com.walletconnect.sample.common.ui.coloredShadow
 import com.walletconnect.sample.common.ui.commons.BlueButton
 import com.walletconnect.sample.common.ui.conditionalModifier
 import com.walletconnect.sample.common.ui.theme.PreviewTheme
+import com.walletconnect.sample.common.ui.themedColor
 import com.walletconnect.sample.common.ui.toColor
 import com.walletconnect.sample.dapp.BuildConfig
 import com.walletconnect.sample.dapp.ui.DappSampleEvents
 import com.walletconnect.sample.dapp.ui.routes.Route
 import com.walletconnect.sample.dapp.ui.routes.bottom_routes.PairingSelectionResult
 import com.walletconnect.sample.dapp.ui.routes.bottom_routes.pairingSelectionResultKey
-import com.walletconnect.sample.dapp.ui.routes.host.DappSampleViewModel
 import com.walletconnect.wcmodal.client.WalletConnectModal
 import com.walletconnect.wcmodal.ui.openWalletConnectModal
 import com.walletconnect.wcmodal.ui.state.rememberModalState
 import kotlinx.coroutines.flow.collectLatest
 
 @Composable
-fun ChainSelectionRoute(navController: NavController, dappViewModel: DappSampleViewModel) {
-    val composableScope = rememberCoroutineScope()
+fun ChainSelectionRoute(navController: NavController) {
     val context = LocalContext.current
     val viewModel: ChainSelectionViewModel = viewModel()
     val chainsState by viewModel.uiState.collectAsState()
     val isModalState = rememberModalState(navController = navController)
     val isOpen by isModalState.isOpenFlow.collectAsState(initial = isModalState.isOpen)
+    val awaitingProposalResponse = viewModel.awaitingSharedFlow.collectAsState(false).value
 
     LaunchedEffect(Unit) {
         navController.currentBackStackEntryFlow.collectLatest { event ->
@@ -88,11 +97,12 @@ fun ChainSelectionRoute(navController: NavController, dappViewModel: DappSampleV
         viewModel.walletEvents.collect { event ->
             when (event) {
                 DappSampleEvents.SessionApproved -> {
+                    viewModel.awaitingProposalResponse(false)
                     navController.navigate(Route.Session.path)
                 }
 
                 DappSampleEvents.SessionRejected -> {
-//                    dappViewModel.awaitingProposalResponse(false)
+                    viewModel.awaitingProposalResponse(false)
                     Toast.makeText(context, "Session has been rejected", Toast.LENGTH_SHORT).show()
                 }
 
@@ -103,6 +113,7 @@ fun ChainSelectionRoute(navController: NavController, dappViewModel: DappSampleV
 
     ChainSelectionScreen(
         chains = chainsState,
+        awaitingState = awaitingProposalResponse,
         isSampleWalletInstalled = context.isSampleWalletInstalled(),
         onChainClick = viewModel::updateChainSelectState,
         onConnectClick = {
@@ -118,62 +129,98 @@ fun ChainSelectionRoute(navController: NavController, dappViewModel: DappSampleV
             } else {
                 Toast.makeText(context, "Please select a chain", Toast.LENGTH_SHORT).show()
             }
-        },
-        onConnectSampleWalletClick = {
-            if (viewModel.isAnyChainSelected) {
-//                dappViewModel.awaitingProposalResponse(true)
-                viewModel.connectToWallet(
-                    onSuccess = { uri ->
-                        val intent = Intent(Intent.ACTION_VIEW).apply {
-                            data = uri.replace("wc:", "wc://").toUri()
-                            `package` = when (BuildConfig.BUILD_TYPE) {
-                                "debug" -> SAMPLE_WALLET_DEBUG_PACKAGE
-                                "internal" -> SAMPLE_WALLET_INTERNAL_PACKAGE
-                                else -> SAMPLE_WALLET_RELEASE_PACKAGE
-                            }
-                        }
-                        context.startActivity(intent)
-                    },
-                    onError = { error ->
-                        dappViewModel.awaitingProposalResponse(false)
-                        Toast.makeText(context, "Error while connecting: $error", Toast.LENGTH_SHORT).show()
-                    })
-            } else {
-                Toast.makeText(context, "Please select a chain", Toast.LENGTH_SHORT).show()
-            }
         }
-    )
+    ) {
+        if (viewModel.isAnyChainSelected) {
+            viewModel.awaitingProposalResponse(true)
+            viewModel.connectToWallet(
+                onSuccess = { uri ->
+                    val intent = Intent(Intent.ACTION_VIEW).apply {
+                        data = uri.replace("wc:", "wc://").toUri()
+                        `package` = when (BuildConfig.BUILD_TYPE) {
+                            "debug" -> SAMPLE_WALLET_DEBUG_PACKAGE
+                            "internal" -> SAMPLE_WALLET_INTERNAL_PACKAGE
+                            else -> SAMPLE_WALLET_RELEASE_PACKAGE
+                        }
+                    }
+                    context.startActivity(intent)
+                },
+                onError = { error ->
+                    viewModel.awaitingProposalResponse(false)
+                    Toast.makeText(context, "Error while connecting: $error", Toast.LENGTH_SHORT).show()
+                })
+        } else {
+            Toast.makeText(context, "Please select a chain", Toast.LENGTH_SHORT).show()
+        }
+    }
 }
 
 @Composable
 private fun ChainSelectionScreen(
     chains: List<ChainSelectionUi>,
+    awaitingState: Boolean,
     isSampleWalletInstalled: Boolean,
     onChainClick: (Int, Boolean) -> Unit,
     onConnectClick: () -> Unit,
     onConnectSampleWalletClick: () -> Unit,
 ) {
-    Column(modifier = Modifier.fillMaxSize()) {
-        WCTopAppBarLegacy(titleText = "Chain selection")
-        ChainsList(
-            chains = chains,
-            modifier = Modifier
-                .weight(1f)
-                .padding(horizontal = 16.dp),
-            onChainClick,
-        )
-        BlueButton(
-            text = "Connect via WalletConnect Modal",
-            onClick = onConnectClick,
-            modifier = Modifier
-                .padding(vertical = 10.dp)
-                .fillMaxWidth()
-                .height(50.dp)
-                .padding(horizontal = 16.dp),
-        )
-        if (isSampleWalletInstalled) {
-            WalletConnectSampleButton(onConnectSampleWalletClick)
+
+
+    Box {
+        Column(modifier = Modifier.fillMaxSize()) {
+            WCTopAppBarLegacy(titleText = "Chain selection")
+            ChainsList(
+                chains = chains,
+                modifier = Modifier
+                    .weight(1f)
+                    .padding(horizontal = 16.dp),
+                onChainClick,
+            )
+            BlueButton(
+                text = "Connect via WalletConnect Modal",
+                onClick = onConnectClick,
+                modifier = Modifier
+                    .padding(vertical = 10.dp)
+                    .fillMaxWidth()
+                    .height(50.dp)
+                    .padding(horizontal = 16.dp),
+            )
+            if (isSampleWalletInstalled) {
+                WalletConnectSampleButton(onConnectSampleWalletClick)
+            }
         }
+        if (awaitingState) {
+            Loader()
+        }
+    }
+}
+
+@Composable
+private fun BoxScope.Loader() {
+    Column(
+        modifier = Modifier
+            .align(Alignment.Center)
+            .clip(RoundedCornerShape(34.dp))
+            .background(themedColor(Color(0xFF242425).copy(alpha = .95f), Color(0xFFF2F2F7).copy(alpha = .95f)))
+            .padding(24.dp),
+        verticalArrangement = Arrangement.Center,
+        horizontalAlignment = Alignment.CenterHorizontally,
+    ) {
+        CircularProgressIndicator(
+            strokeWidth = 8.dp,
+            modifier = Modifier
+                .size(75.dp), color = Color(0xFFB8F53D)
+        )
+        Spacer(modifier = Modifier.height(16.dp))
+        Text(
+            text = "Awaiting response...",
+            maxLines = 1,
+            style = TextStyle(
+                fontWeight = FontWeight.Medium,
+                fontSize = 22.sp,
+                color = themedColor(Color(0xFFb9b3b5), Color(0xFF484648))
+            ),
+        )
     }
 }
 
@@ -250,11 +297,11 @@ private fun ChainSelectionScreenPreview(
     PreviewTheme {
         ChainSelectionScreen(
             chains = chains,
+            awaitingState = false,
             isSampleWalletInstalled = false,
             onChainClick = { _, _ -> },
-            onConnectClick = {},
-            onConnectSampleWalletClick = {}
-        )
+            onConnectClick = {}
+        ) {}
     }
 }
